Adds support for optional "cmyk" parameter when outputting EPS.
This adds a new, optional argument to the eps function, which
allows you to use CMYK colors instead of the regular RGB when
creating the EPS file.

This allows you to generate QR codes with true black instead of
the automagic four color black default when using the QR code in
a CMYK based medium such as the newspaper or printing industry.

Example:

QRcode::eps($text, false, QR_ECLEVEL_L, 40, 4, false, 0x00000000, 0x000000FF, true);

The back / front colors still mean the same, but we now also
use the upper byte (for the C value).